Creating separate component for `ActionDropdown` to ensure we are only iterating through all field and value actions when opening action dropdown.
